+++ backend/src/main/java/com/storerader/server/domain/price/dto/PriceItemDTO.java
@@ -1,7 +1,7 @@
 package com.storerader.server.domain.price.dto;
 
 public record PriceItemDTO(
-        Integer storeId,
+        Long storeId,
         Integer price,
         String inspectDay
 ) {
